Fix a leak of RTCVideoView and RTCVideoTrack
RTCVideoView strongly retains the RTCVideoTrack it renders via its
property currentRenderer. RTCVideoTrack strongly retains the
RTCVideoView which the former sees as its RTCVideoRenderer via
the method addRenderer:. This is a cycle of strong retainments and,
consequently, it causes the memory associated with at least RTCVideoView
and RTCVideoTrack to leak (after the two instances are no longer
reachable by the application). The leak is observed to increase the
app's memory usage over time until the process' memory is exhausted and
the app crashes.

The fix of the memory leak additionally fixes a bug in RTCVideoView
which allowed multiple RTCVideoTracks to be rendered into it.