Default RTCPeerConnection constraints in JavaScript
The default RTCPeerConnection constraints passed inside peerConnectionInit,
createOffer(), and createAnswer() were the same on Android and iOS but
were defined and passed in the Java and Objective-C source code,
respectively. Which means that (1) the source code was duplicated and,
consequently, (2) the risk of mistakenly diverging was higher and (3)
maintenance was hampered because adding a new default constraint needed
duplicate work. Move the default RTCPeerConnection definitions in the
JavaScript source code to have them in one place and to make it easier
to add new key-value pairs.