android: enable HW acceleation for video
This patch may seem non-obvious. It all has to do with having a root EGL
base object of which the rest EGL contexts and bases are created.

EGL contexts are needed in 3 places and they all need to be related to
the same root:

- factory.setVideoHwAccelerationOptions
- factory.createVideoSource
- SurfaceViewRenderer.init

The approach taken here is to have a global EGL root which never gets
released and pass it around as needed.